Migrated from mapStruct to modelMapper, bug fix

+++ src/main/java/com/rostikandrusiv/epamlab29/spring/mvc/rest/repository/MovieRepository.java
@@ -9,6 +9,6 @@ public interface MovieRepository extends JpaRepository<Movie, Long> {
 
     Optional<Movie> findByName(String movieName);
 
-    Optional<Movie> findById(int id);
+    Optional<Movie> findById(long id);
 
 }
